+++ src/playback/view/toolbars/playbacktoolbarlevelitem.h
@@ -17,7 +17,7 @@ class PlaybackToolBarLevelItem : public muse::uicomponents::ToolBarItem
 {
     Q_OBJECT
 
-    Q_PROPERTY(int level READ level WRITE setLevel NOTIFY levelChanged FINAL)
+    Q_PROPERTY(float level READ level WRITE setLevel NOTIFY levelChanged FINAL)
 
     Q_PROPERTY(float leftChannelPressure READ leftChannelPressure NOTIFY leftChannelPressureChanged)
     Q_PROPERTY(float leftChannelRMS READ leftChannelRMS NOTIFY leftChannelRMSChanged)
@@ -45,8 +45,8 @@ public:
     explicit PlaybackToolBarLevelItem(const muse::ui::UiAction& action, muse::uicomponents::ToolBarItemType::Type type,
                                       QObject* parent = nullptr);
 
-    int level() const;
-    void setLevel(int newLevel);
+    float level() const;
+    void setLevel(float newLevel);
 
     float leftChannelPressure() const;
     float leftChannelRMS() const;
@@ -106,7 +106,7 @@ private:
     void setAudioChannelRMS(const audio::audioch_t chNum, const float newValue);
     void resetAudioChannelsVolumePressure();
 
-    int m_level = 0;
+    float m_level = 0;
 
     float m_leftChannelPressure = -60.0;
     float m_leftChannelRMS = -60.0;
